feat: Move useSearchQueryState to dataset-search
